Update RubicCube.py
test_complimentaryOperation: This function tests one operation produces the same result as the associated operation once the cube has been rotated.
    For instance, rotatingLeftClockwise should produce the same result as rotating the cube from left to right, rotating the front face clockwise, and rotating the cube back from right to left

And with this one. I think I have finished with all the interesting test functions for the cube operations. So my next steps is writting the skeleton of the search algorithm!